add MN, wallet update, general update

+++ BitcoinGenX_5pack.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#0.9g-- NullEntryDev Script
+#0.99-- NullEntryDev Script
 NODESL=Five
 NODESN=5
 BLUE='\033[0;96m'
@@ -27,11 +27,51 @@ echo -e ${BLUE}"Zero Confidental information or Wallet keys will be stored in it
 echo -e ${YELLOW}"Press y to agree followed by [ENTER], or just [ENTER] to disagree"${CLEAR}
 read NULLREC
 echo
+echo -e ${GREEN}"Would you like to enter custom IP addresses?"${CLEAR}
+echo -e ${YELLOW}"If you don't know the answer, hit n for no"${CLEAR}
+echo -e ${YELLOW}"If you have custom IPs hit y for yes"${CLEAR}
+read customIP
+echo "Creating ${NODESN} BitcoinGenX system user(s) with no-login access:"
+if id "bitcoingenx" >/dev/null 2>&1; then
+echo "user exists"
+MN1=1
+else
+sudo adduser --system --home /home/bitcoingenx bitcoingenx
+MN1=0
+fi
+if id "bitcoingenx2" >/dev/null 2>&1; then
+echo -e ${YELLOW} "Found user bitcoingenx2!"${CLEAR}
+MN2=1
+else
+sudo adduser --system --home /home/bitcoingenx2 bitcoingenx2
+MN2=0
+fi
+if id "bitcoingenx3" >/dev/null 2>&1; then
+echo -e ${YELLOW} "Found user bitcoingenx3!"${CLEAR}
+MN3=1
+else
+sudo adduser --system --home /home/bitcoingenx3 bitcoingenx3
+MN3=0
+fi
+if id "bitcoingenx4" >/dev/null 2>&1; then
+echo -e ${YELLOW} "Found user bitcoingenx4!"${CLEAR}
+MN4=1
+else
+sudo adduser --system --home /home/bitcoingenx4 bitcoingenx4
+MN4=0
+fi
+if id "bitcoingenx5" >/dev/null 2>&1; then
+echo -e ${YELLOW} "Found user bitcoingenx5!"${CLEAR}
+MN5=1
+else
+sudo adduser --system --home /home/bitcoingenx5 bitcoingenx5
+MN5=0
+fi
 echo
 echo
 echo
 echo
-echo -e ${RED}"Your Masternode Private Keys are needed,"${CLEAR}
+echo -e ${RED}"Your New Masternode Private Keys are needed,"${CLEAR}
 echo -e ${GREEN}" -which can be generated from the local wallet"${CLEAR}
 echo
 echo -e ${YELLOW}"You can edit the config later if you don't have this"${CLEAR}
@@ -40,27 +80,41 @@ echo -e ${YELLOW}"And the script installation will hang and fail"${CLEAR}
 echo
 echo -e ${YELLOW}"Right Click to paste in some SSH Clients"${CLEAR}
 echo
+if [[ "$MN1" -eq "0" ]]; then
 echo -e ${GREEN}"Please Enter Your First Masternode Private Key:"${CLEAR}
-read privkey
+read MNKEY
 echo
+else
+echo -e ${YELLOW}"Skipping First Masternode Key"${CLEAR}
+fi
+if [[ "$MN2" -eq "0" ]]; then
 echo -e ${GREEN}"Please Enter Your Second Masternode Private Key:"${CLEAR}
-read privkey2
+read MNKEY2
 echo
+else
+echo -e ${YELLOW}"Skipping Second Masternode Key"${CLEAR}
+fi
+if [[ "$MN3" -eq "0" ]]; then
 echo -e ${GREEN}"Please Enter Your Third Masternode Private Key:"${CLEAR}
-read privkey3
+read MNKEY3
 echo
+else
+echo -e ${YELLOW}"Skipping Third Masternode Key"${CLEAR}
+fi
+if [[ "$MN4" -eq "0" ]]; then
 echo -e ${GREEN}"Please Enter Your Fourth Masternode Private Key:"${CLEAR}
-read privkey4
+read MNKEY4
 echo
+else
+echo -e ${YELLOW}"Skipping Fourth Masternode Key"${CLEAR}
+fi
+if [[ "$MN5" -eq "0" ]]; then
 echo -e ${GREEN}"Please Enter Your Fifth Masternode Private Key:"${CLEAR}
-read privkey5
+read MNKEY5
 echo
-echo "Creating ${NODESN} BitcoinGenX system users with no-login access:"
-sudo adduser --system --home /home/bitcoingenx bitcoingenx
-sudo adduser --system --home /home/bitcoingenx2 bitcoingenx2
-sudo adduser --system --home /home/bitcoingenx3 bitcoingenx3
-sudo adduser --system --home /home/bitcoingenx4 bitcoingenx4
-sudo adduser --system --home /home/bitcoingenx5 bitcoingenx5
+else
+echo -e ${YELLOW}"Skipping Fifth Masternode Key"${CLEAR}
+fi
 cd ~
 if [[ $NULLREC = "y" ]] ; then
 if [ ! -d /usr/local/nullentrydev/ ]; then
@@ -113,6 +167,18 @@ if [[ $NULLREC = "y" ]] ; then
 echo "dependenciesInstalled: true" >> /usr/local/nullentrydev/mnodes.log
 fi
 fi
+if [[ customIP = "y" ]] ; then
+echo -e ${GREEN}"IP for Masternode 1"${CLEAR}
+read MNIP1
+echo -e ${GREEN}"IP for Masternode 2"${CLEAR}
+read MNIP2
+echo -e ${GREEN}"IP for Masternode 3"${CLEAR}
+read MNIP3
+echo -e ${GREEN}"IP for Masternode 4"${CLEAR}
+read MNIP4
+echo -e ${GREEN}"IP for Masternode 5"${CLEAR}
+read MNIP5
+else
 regex='^([0-9a-fA-F]{0,4}:){1,7}[0-9a-fA-F]{0,4}$'
 FINDIP=$(hostname -I | cut -f2 -d' '| cut -f1-7 -d:)
 if [[ $FINDIP =~ $regex ]]; then
@@ -149,29 +215,92 @@ MNIP2=$(sed -n '2p' < ip.tmp)
 MNIP3=$(sed -n '3p' < ip.tmp)
 MNIP4=$(sed -n '4p' < ip.tmp)
 MNIP5=$(sed -n '5p' < ip.tmp)
-if [[ $NULLREC = "y" ]] ; then
-sudo touch /usr/local/nullentrydev/iptable.log
-sudo cp ip.tmp >> /usr/local/nullentrydev/iptable.log
-fi
 rm -rf ip.tmp
+fi
+if grep -Fxq "swapInstalled: true" /usr/local/nullentrydev/mnodes.log
+then
+echo -e ${RED}"Skipping... Swap Area already made"${CLEAR}
+else
 cd /var
 sudo touch swap.img
 sudo chmod 600 swap.img
 sudo dd if=/dev/zero of=/var/swap.img bs=1024k count=4096
 sudo mkswap /var/swap.img
 sudo swapon /var/swap.img
+if [[ $NULLREC = "y" ]] ; then
+echo "swapInstalled: true" >> /usr/local/nullentrydev/mnodes.log
+fi
+fi
 cd ~
+touch bgxcheck.tmp
+ps aux | grep bitcoingenx >> bgxcheck.tmp
+if grep home/bitcoingenx/.bitcoingenx bgxcheck.tmp
+then
+echo Found OLD ${NC} bgx Node running
+OldNode="1"
+else
+echo No ${NC} bgx Node not running
+OldNode="0"
+fi
+until [[ $NC = 9 ]]; do
+if grep /home/bitcoingenx${NC}/.bitcoingenx bgxcheck.tmp
+then
+echo Found ${NC} bgx Node running
+declare IPN$NC="1"
+RB=1
+else
+echo No ${NC} bgx Node not running
+declare IPN$NC="0"
+echo $NC
+fi
+NC=$[$NC+1]
+done
+rm -r bgxcheck.tmp
+if [[ "$OldNode" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx/.bitcoingenx stop
+fi
+if [[ "$IPN1" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx1/.bitcoingenx stop
+fi
+if [[ "$IPN2" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx2/.bitcoingenx stop
+fi
+if [[ "$IPN3" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx3/.bitcoingenx stop
+fi
+if [[ "$IPN4" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx4/.bitcoingenx stop
+fi
+if [[ "$IPN5" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx5/.bitcoingenx stop
+fi
+if [[ "$IPN6" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx6/.bitcoingenx stop
+fi
+if [[ "$IPN7" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx7/.bitcoingenx stop
+fi
+if [[ "$IPN8" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx8/.bitcoingenx stop
+fi
+if [[ "$IPN9" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx9/.bitcoingenx stop
+fi
+if [[ "$IPN0" = "1" ]]; then
+bitcoingenx-cli -datadir=/home/bitcoingenx0/.bitcoingenx stop
+fi
 if [ ! -d /root/bgx ]; then
 sudo mkdir /root/bgx
 fi
 cd /root/bgx
 echo "Downloading latest BitcoinGenX binaries"
-wget https://github.com/BitcoinGenX/BitcoinGenesisX/files/2853315/bitcoingenx-linux.zip
-unzip bitcoingenx-linux.zip
+wget https://github.com/BitcoinGenX/BitcoinGenesisX/files/2896837/bitcoingenx-linux-static.zip
+unzip bitcoingenx-linux-static.zip
 sleep 3
 sudo mv /root/bgx/bitcoingenxd /root/bgx/bitcoingenx-cli /usr/local/bin
 sudo chmod 755 -R /usr/local/bin/bitcoingenx*
 rm -rf /root/bgx
+if [ ! -f /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf ]; then
 echo -e "${GREEN}Configuring First BitcoinGenX Node${CLEAR}"
 sudo mkdir /home/bitcoingenx/.bitcoingenx
 sudo touch /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
@@ -185,15 +314,20 @@ echo "masternode=1" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
 echo "rpcport=19021" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
 echo "listen=0" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
 echo "externalip=[${MNIP1}]:4488" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
-echo "masternodeprivkey=$privkey" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
+echo "masternodeprivkey=$MNKEY" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
+echo "addnode=23.94.102.195:4488" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
+MN1=0
 if [[ $NULLREC = "y" ]] ; then
 echo "masterNode1 : true" >> /usr/local/nullentrydev/bgx.log
 echo "walletVersion1 : 1.4.0COINVERSION=1.6.0" >> /usr/local/nullentrydev/bgx.log
-echo "scriptVersion1 : 0.9g" >> /usr/local/nullentrydev/bgx.log
+echo "scriptVersion1 : 0.99" >> /usr/local/nullentrydev/bgx.log
 fi
-sleep 5
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Configuration there"${CLEAR}
+fi
 echo
-echo -e ${YELLOW}"Launching first BGX Node"${CLEAR}
+echo -e ${YELLOW}"Launching First BGX Node"${CLEAR}
 bitcoingenxd -datadir=/home/bitcoingenx/.bitcoingenx -daemon
 echo
 echo -e ${YELLOW}"Looking for a Shared Masternode Service? Check out Crypto Hash Tank" ${CLEAR}
@@ -202,7 +336,10 @@ echo -e ${YELLOW}" https://www.cryptohashtank.com/TJIF "${CLEAR}
 echo
 echo -e ${YELLOW}"Special Thanks to the BitcoinGenX (BGX) Community" ${CLEAR}
 sleep 20
-echo -e "${GREEN}Configuring second BitcoinGenX Node${CLEAR}"
+if [ ! -f /home/bitcoingenx2/.bitcoingenx/bitcoingenx.conf ]; then
+if [ ! -f /home/bitcoingenx2/bitcoingenx.conf ]; then
+echo -e "${YELLOW}Second BitcoinGenX Normal Warning - Node Configuration Not Found....${CLEAR}"
+echo -e "${GREEN}Configuring Second BitcoinGenX Node${CLEAR}"
 sudo mkdir /home/bitcoingenx2/.bitcoingenx
 sudo touch /home/bitcoingenx2/bitcoingenx.conf
 echo "rpcuser=user"`shuf -i 100000-9999999 -n 1` >> /home/bitcoingenx2/bitcoingenx.conf
@@ -215,15 +352,28 @@ echo "masternode=1" >> /home/bitcoingenx2/bitcoingenx.conf
 echo "rpcport=19022" >> /home/bitcoingenx2/bitcoingenx.conf
 echo "listen=0" >> /home/bitcoingenx2/bitcoingenx.conf
 echo "externalip=[${MNIP2}]:4488" >> /home/bitcoingenx2/bitcoingenx.conf
-echo "masternodeprivkey=$privkey2" >> /home/bitcoingenx2/bitcoingenx.conf
+echo "masternodeprivkey=$MNKEY2" >> /home/bitcoingenx2/bitcoingenx.conf
+echo "addnode=[${MNIP1}]" >> /home/bitcoingenx/.bitcoingenx/bitcoingenx.conf
 if [[ $NULLREC = "y" ]] ; then
 echo "masterNode2 : true" >> /usr/local/nullentrydev/bgx.log
 echo "walletVersion2 : 1.4.0COINVERSION=1.6.0" >> /usr/local/nullentrydev/bgx.log
-echo "scriptVersion2 : 0.9g" >> /usr/local/nullentrydev/bgx.log
+echo "scriptVersion2 : 0.99" >> /usr/local/nullentrydev/bgx.log
 fi
-sleep 5
+else
 echo
-echo -e "${GREEN}Configuring third BitcoinGenX Node${CLEAR}"
+echo -e ${GREEN}"Found /home/bitcoingenx2/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Pre-stage for Second Node "${CLEAR}
+MN2=0
+fi
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx2/.bitcoingenx/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Configuration for Second Node"${CLEAR}
+fi
+echo
+if [ ! -f /home/bitcoingenx3/.bitcoingenx/bitcoingenx.conf ]; then
+if [ ! -f /home/bitcoingenx3/bitcoingenx.conf ]; then
+echo -e "${GREEN}Third BitcoinGenX Node Configuration Not Found....${CLEAR}"
+echo -e "${GREEN}Configuring Third BitcoinGenX Node${CLEAR}"
 sudo mkdir /home/bitcoingenx3/.bitcoingenx
 sudo touch /home/bitcoingenx3/bitcoingenx.conf
 echo "rpcuser=user"`shuf -i 100000-9999999 -n 1` >> /home/bitcoingenx3/bitcoingenx.conf
@@ -233,18 +383,30 @@ echo "server=1" >> /home/bitcoingenx3/bitcoingenx.conf
 echo "daemon=1" >> /home/bitcoingenx3/bitcoingenx.conf
 echo "maxconnections=250" >> /home/bitcoingenx3/bitcoingenx.conf
 echo "masternode=1" >> /home/bitcoingenx3/bitcoingenx.conf
-echo "rpcport=19023" >> /home/bitcoingenx3/bitcoingenx.conf
+echo "rpcport=19024" >> /home/bitcoingenx3/bitcoingenx.conf
 echo "listen=0" >> /home/bitcoingenx3/bitcoingenx.conf
 echo "externalip=[${MNIP3}]:4488" >> /home/bitcoingenx3/bitcoingenx.conf
-echo "masternodeprivkey=$privkey3" >> /home/bitcoingenx3/bitcoingenx.conf
+echo "masternodeprivkey=$MNKEY3" >> /home/bitcoingenx3/bitcoingenx.conf
 if [[ $NULLREC = "y" ]] ; then
 echo "masterNode3 : true" >> /usr/local/nullentrydev/bgx.log
 echo "walletVersion3 : 1.4.0COINVERSION=1.6.0" >> /usr/local/nullentrydev/bgx.log
-echo "scriptVersion3 : 0.9g" >> /usr/local/nullentrydev/bgx.log
+echo "scriptVersion3 : 0.99" >> /usr/local/nullentrydev/bgx.log
 fi
-sleep 5
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx3/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Pre-stage for Third Node "${CLEAR}
+MN3=0
+fi
 echo
-echo -e "${GREEN}Configuring fourth BitcoinGenX Node${CLEAR}"
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx3/.bitcoingenx/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Configuration for Third Node"${CLEAR}
+fi
+echo
+if [ ! -f /home/bitcoingenx4/.bitcoingenx/bitcoingenx.conf ]; then
+if [ ! -f /home/bitcoingenx4/bitcoingenx.conf ]; then
+echo -e "${GREEN}Fourth BitcoinGenX Node Configuration Not Found....${CLEAR}"
+echo -e "${GREEN}Configuring Fourth BitcoinGenX Node${CLEAR}"
 sudo mkdir /home/bitcoingenx4/.bitcoingenx
 sudo touch /home/bitcoingenx4/bitcoingenx.conf
 echo "rpcuser=user"`shuf -i 100000-9999999 -n 1` >> /home/bitcoingenx4/bitcoingenx.conf
@@ -254,18 +416,29 @@ echo "server=1" >> /home/bitcoingenx4/bitcoingenx.conf
 echo "daemon=1" >> /home/bitcoingenx4/bitcoingenx.conf
 echo "maxconnections=250" >> /home/bitcoingenx4/bitcoingenx.conf
 echo "masternode=1" >> /home/bitcoingenx4/bitcoingenx.conf
-echo "rpcport=19024" >> /home/bitcoingenx4/bitcoingenx.conf
+echo "rpcport=19025" >> /home/bitcoingenx4/bitcoingenx.conf
 echo "listen=0" >> /home/bitcoingenx4/bitcoingenx.conf
 echo "externalip=[${MNIP4}]:4488" >> /home/bitcoingenx4/bitcoingenx.conf
-echo "masternodeprivkey=$privkey4" >> /home/bitcoingenx4/bitcoingenx.conf
+echo "masternodeprivkey=$MNKEY4" >> /home/bitcoingenx4/bitcoingenx.conf
 if [[ $NULLREC = "y" ]] ; then
 echo "masterNode4 : true" >> /usr/local/nullentrydev/bgx.log
 echo "walletVersion4 : 1.4.0COINVERSION=1.6.0" >> /usr/local/nullentrydev/bgx.log
-echo "scriptVersion4 : 0.9g" >> /usr/local/nullentrydev/bgx.log
+echo "scriptVersion4 : 0.99" >> /usr/local/nullentrydev/bgx.log
 fi
-sleep 5
+else
 echo
-echo -e "${GREEN}Configuring fifth BitcoinGenX Node${CLEAR}"
+echo -e ${YELLOW}"Found /home/bitcoingenx4/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Pre-stage for Fourth Node "${CLEAR}
+MN4=0
+fi
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx4/.bitcoingenx/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Configuration for Fourth Node"${CLEAR}
+fi
+if [ ! -f /home/bitcoingenx5/.bitcoingenx/bitcoingenx.conf ]; then
+if [ ! -f /home/bitcoingenx5/bitcoingenx.conf ]; then
+echo -e "${GREEN}Fifth BitcoinGenX Node Configuration Not Found....${CLEAR}"
+echo -e "${GREEN}Configuring Fifth BitcoinGenX Node${CLEAR}"
 sudo mkdir /home/bitcoingenx5/.bitcoingenx
 sudo touch /home/bitcoingenx5/bitcoingenx.conf
 echo "rpcuser=user"`shuf -i 100000-9999999 -n 1` >> /home/bitcoingenx5/bitcoingenx.conf
@@ -275,17 +448,25 @@ echo "server=1" >> /home/bitcoingenx5/bitcoingenx.conf
 echo "daemon=1" >> /home/bitcoingenx5/bitcoingenx.conf
 echo "maxconnections=250" >> /home/bitcoingenx5/bitcoingenx.conf
 echo "masternode=1" >> /home/bitcoingenx5/bitcoingenx.conf
-echo "rpcport=19025" >> /home/bitcoingenx5/bitcoingenx.conf
+echo "rpcport=19026" >> /home/bitcoingenx5/bitcoingenx.conf
 echo "listen=0" >> /home/bitcoingenx5/bitcoingenx.conf
 echo "externalip=[${MNIP5}]:4488" >> /home/bitcoingenx5/bitcoingenx.conf
-echo "masternodeprivkey=$privkey5" >> /home/bitcoingenx5/bitcoingenx.conf
+echo "masternodeprivkey=$MNKEY5" >> /home/bitcoingenx5/bitcoingenx.conf
 if [[ $NULLREC = "y" ]] ; then
 echo "masterNode5 : true" >> /usr/local/nullentrydev/bgx.log
 echo "walletVersion5 : 1.4.0COINVERSION=1.6.0" >> /usr/local/nullentrydev/bgx.log
-echo "scriptVersion5 : 0.9g" >> /usr/local/nullentrydev/bgx.log
+echo "scriptVersion5 : 0.99" >> /usr/local/nullentrydev/bgx.log
 fi
-sleep 5
+else
 echo
+echo -e ${YELLOW}"Found /home/bitcoingenx5/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Pre-stage for Fifth Node "${CLEAR}
+MN5=0
+fi
+else
+echo -e ${YELLOW}"Found /home/bitcoingenx5/.bitcoingenx/bitcoingenx.conf"${CLEAR}
+echo -e ${YELLOW}"Skipping Configuration for Fifth Node"${CLEAR}
+fi
 echo -e "${RED}This process can take a while!${CLEAR}"
 echo -e "${YELLOW}Waiting on First Masternode Block Chain to Synchronize${CLEAR}"
 echo -e "${YELLOW}Once complete, it will stop and copy the block chain to${CLEAR}"
@@ -300,30 +481,30 @@ echo -e "${GREEN}Haulting and Replicating First BitcoinGenX Node${CLEAR}"
 
 bitcoingenx-cli -datadir=/home/bitcoingenx/.bitcoingenx stop
 sleep 10
+if [[ "$MN2" -eq "0" ]]; then
 sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx2/.bitcoingenx
-sleep 3
-sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx3/.bitcoingenx
-sleep 3
-sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx4/.bitcoingenx
-sleep 3
-sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx5/.bitcoingenx
-sleep 3
 rm /home/bitcoingenx2/.bitcoingenx/bitcoingenx.conf
-sleep 1
-rm /home/bitcoingenx3/.bitcoingenx/bitcoingenx.conf
-sleep 1
-rm /home/bitcoingenx4/.bitcoingenx/bitcoingenx.conf
-sleep 1
-rm /home/bitcoingenx5/.bitcoingenx/bitcoingenx.conf
-sleep 1
 cp -r /home/bitcoingenx2/bitcoingenx.conf /home/bitcoingenx2/.bitcoingenx/bitcoingenx.conf
 sleep 1
+fi
+if [[ "$MN3" -eq "0" ]]; then
+sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx3/.bitcoingenx
+rm /home/bitcoingenx3/.bitcoingenx/bitcoingenx.conf
 cp -r /home/bitcoingenx3/bitcoingenx.conf /home/bitcoingenx3/.bitcoingenx/bitcoingenx.conf
 sleep 1
+fi
+if [[ "$MN4" -eq "0" ]]; then
+sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx4/.bitcoingenx
+rm /home/bitcoingenx4/.bitcoingenx/bitcoingenx.conf
 cp -r /home/bitcoingenx4/bitcoingenx.conf /home/bitcoingenx4/.bitcoingenx/bitcoingenx.conf
 sleep 1
+fi
+if [[ "$MN5" -eq "0" ]]; then
+sudo cp -r /home/bitcoingenx/.bitcoingenx/* /home/bitcoingenx5/.bitcoingenx
+rm /home/bitcoingenx5/.bitcoingenx/bitcoingenx.conf
 cp -r /home/bitcoingenx5/bitcoingenx.conf /home/bitcoingenx5/.bitcoingenx/bitcoingenx.conf
 sleep 1
+fi
 echo -e ${YELLOW}"Launching First BGX Node"${CLEAR}
 bitcoingenxd -datadir=/home/bitcoingenx/.bitcoingenx -daemon
 sleep 20
